refactor(cli): extract useImageSupportProbe hook and dedupe dialog test scaffolding
